Refactored. Media type processing dependency on access facet is clearer and less buried now
